Moved FactFilterDsl to its own package
