+++ core/src/main/scala/com/rallyhealth/vapors/core/dsl/factfilter/LogicalOps.scala
@@ -1,6 +1,7 @@
-package com.rallyhealth.vapors.core.dsl
+package com.rallyhealth.vapors.core.dsl.factfilter
 
 import com.rallyhealth.vapors.core.dsl
+import com.rallyhealth.vapors.core.dsl.AnyExp
 import com.rallyhealth.vapors.core.logic.{Intersect, Union}
 
 final class LogicalOps[T, A](private val exp: AnyExp[T, A]) extends AnyVal {
