+++ core/src/main/scala/com/rallyhealth/vapors/core/dsl/factfilter/TypedFactExpBuilder.scala
@@ -1,6 +1,7 @@
-package com.rallyhealth.vapors.core.dsl
+package com.rallyhealth.vapors.core.dsl.factfilter
 
 import com.rallyhealth.vapors.core.data.{Fact, FactTypeSet, NamedLens}
+import com.rallyhealth.vapors.core.dsl.{CondExp, TerminalFactsExp}
 
 import scala.reflect.ClassTag
 import scala.reflect.runtime.universe.TypeTag
