+++ core/src/main/scala/com/rallyhealth/vapors/core/dsl/factfilter/package.scala
@@ -1,0 +1,3 @@
+package com.rallyhealth.vapors.core.dsl
+
+package object factfilter extends Dsl
