+++ core/src/main/scala/com/rallyhealth/vapors/core/dsl/package.scala
@@ -1,3 +1,3 @@
 package com.rallyhealth.vapors.core
 
-package object dsl extends FactFilterDsl
+package object dsl extends factfilter.Dsl
